requirements.txt
# Core dependencies
transformers>=4.45.0
peft>=0.13.0
trl>=0.12.0
datasets>=3.0.0
accelerate>=1.0.0
bitsandbytes>=0.44.0
torch>=2.1.0
sentencepiece>=0.2.0

# Inference
llama-cpp-python>=0.3.0

# Evaluation / Demo
gradio>=5.0.0

# Data generation (optional - only if using Gemini API)
google-generativeai>=0.8.0
